+++ example/lib/main.dart
@@ -18,7 +18,20 @@ import 'package:fusion_example/page/willpop_page.dart';
 void main() {
   print('defaultRouteName=${ui.window.defaultRouteName}');
   Fusion.instance.install();
+  FusionAppLifecycleBinding.instance.register(MyAppLifecycleListener());
   runApp(const MyApp());
+}
+
+class MyAppLifecycleListener implements FusionAppLifecycleListener {
+  @override
+  void onBackground() {
+    print('onBackground');
+  }
+
+  @override
+  void onForeground() {
+    print('onForeground');
+  }
 }
 
 class MyApp extends StatelessWidget {